arm64: support __int128 with clang
Commit fb8722735f50 ("arm64: support __int128 on gcc 5+") added support
for arm64 __int128 with gcc with a version-conditional, but neglected to
enable this for clang, which in fact appears to support aarch64 __int128.
This commit therefore enables it if the compiler is clang, using the
same type of makefile conditional used elsewhere in the tree.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
Change-Id: I24247d985f0c91b3775b7ba16144ca710f59f77c
Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>
(cherry picked from commit 6471732f55dde74c82cf73e92827eecaa72b15ae)
Signed-off-by: KenHV <yo@kenharris.xyz>