msm: kgsl: Remove POPP.
POPP constantly attempts to lower the GPU's frequency behind the
governor's back in order to save power; however, the GPU governor in use
(msm-adreno-tz) is very good at determining the GPU's load and selecting
an appropriate frequency to run the GPU at.

POPP was created long ago, perhaps when msm-adreno-tz didn't exist or
didn't work so well, so it is clearly deprecated. Remove it.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>
Change-Id: I9ee7f8e2c62973dc890739285d45e649f40cf3cd
Signed-off-by: KenHV <yo@kenharris.xyz>