drm: Reduce latency while processing atomic ioctls.
Unfortunately, our display driver must occasionally sleep inside
start_atomic(), which causes the CPU running the ioctl to schedule and
potentially idle. Depending on how deeply the running CPU idles, there
can be quite a bit of latency added to processing the "atomic" ioctl as
a result, which hurts display rendering latency. We can alleviate this
by optimistically assuming that the task running the ioctl won't
migrate CPUs while it runs and restricting the current CPU to shallow
idle states.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>
Change-Id: Ie30f9f2396e11f4c2ee5eff0aa4c7646c13de328
Signed-off-by: KenHV <yo@kenharris.xyz>